apps/web-summarizer/backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0  # This is the generic OpenAI client used by multiple providers
mistralai==0.0.11
google-generativeai==0.4.1
reportlab==4.0.4
python-dotenv==1.0.1